feat
Добавить исключения

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -4,6 +4,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Service;
+import ru.yandex.practicum.filmorate.exeptions.NotFoundException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.model.Genre;
 import ru.yandex.practicum.filmorate.model.Mpa;
@@ -65,8 +66,10 @@ public class FilmService {
 		return filmStorage.getRatingsMpaById(id);
 	}
 
-	public void deleteFilm(int id) {
-		filmStorage.deleteFilm(id);
+	public void deleteFilm(Long id) {
+		if (filmStorage.getFilmById(id) != null) {
+			filmStorage.deleteFilm(id);
+		} else throw new NotFoundException("The user with id does not exist!");
 	}
 
 }
